Rename subfolders to shorten require statements to pass linter
