+++ lib/disability_compensation/factories/api_provider_factory.rb
@@ -18,8 +18,8 @@ require 'disability_compensation/providers/brd/lighthouse_brd_provider'
 require 'disability_compensation/providers/generate_pdf/generate_pdf_provider'
 require 'disability_compensation/providers/generate_pdf/evss_generate_pdf_provider'
 require 'disability_compensation/providers/generate_pdf/lighthouse_generate_pdf_provider'
-require 'disability_compensation/providers/supplemental_document_upload/lighthouse_supplemental_document_upload_provider'
-require 'disability_compensation/providers/supplemental_document_upload/evss_supplemental_document_upload_provider'
+require 'disability_compensation/providers/document_upload/lighthouse_supplemental_document_upload_provider'
+require 'disability_compensation/providers/document_upload/evss_supplemental_document_upload_provider'
 require 'logging/third_party_transaction'
 
 class ApiProviderFactory
